@@ -13,46 +13,47 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "5.2.5",
-    "@angular/common": "5.2.5",
-    "@angular/compiler": "5.2.5",
-    "@angular/core": "5.2.5",
-    "@angular/forms": "5.2.5",
-    "@angular/http": "5.2.5",
-    "@angular/platform-browser": "5.2.5",
-    "@angular/platform-browser-dynamic": "5.2.5",
-    "@angular/router": "5.2.5",
-    "@ng-bootstrap/ng-bootstrap": "^1.0.0",
-    "angularfire2": "^5.0.0-rc.6",
+    "@angular/animations": "6.0.3",
+    "@angular/common": "6.0.3",
+    "@angular/compiler": "6.0.3",
+    "@angular/core": "6.0.3",
+    "@angular/forms": "6.0.3",
+    "@angular/http": "6.0.3",
+    "@angular/platform-browser": "6.0.3",
+    "@angular/platform-browser-dynamic": "6.0.3",
+    "@angular/router": "6.0.3",
+    "@ng-bootstrap/ng-bootstrap": "^2.0.0",
+    "angularfire2": "^5.0.0-rc.10",
     "bootstrap": "^4.0.0",
     "core-js": "^2.5.3",
-    "firebase": "^4.10.0",
+    "firebase": "^5.0.3",
     "font-awesome": "^4.7.0",
     "jquery": "^3.3.1",
     "ng2-validation": "^4.2.0",
     "popper.js": "^1.13.0",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.20"
+    "rxjs": "^6.2.0",
+    "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.0",
-    "@angular/compiler-cli": "5.2.5",
-    "@angular/language-service": "5.2.5",
+    "@angular-devkit/build-angular": "~0.6.3",
+    "@angular/cli": "6.0.3",
+    "@angular/compiler-cli": "6.0.3",
+    "@angular/language-service": "6.0.3",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "^9.4.6",
     "codelyzer": "^4.1.0",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "^2.0.2",
     "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
     "karma-coverage-istanbul-reporter": "^1.4.1",
     "karma-jasmine": "^1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.3.0",
+    "protractor": "^5.3.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "2.6.2"
+    "typescript": "2.7.2"
   }
 }
