@@ -1,6 +1,5 @@
-import { Subscription } from 'rxjs/Subscription';
+import { Subscription ,  Observable } from 'rxjs';
 import { Product } from './../../models/product';
-import { Observable } from 'rxjs/Observable';
 import { ProductService } from './../../services/product.service';
 import { Component, OnInit, OnDestroy } from '@angular/core';
 
