@@ -1,3 +1,5 @@
+
+import {map} from 'rxjs/operators';
 import { Router } from '@angular/router';
 import { AuthService } from './../services/auth.service';
 import { RouterStateSnapshot } from '@angular/router';
@@ -10,9 +12,9 @@ export class AdminAuthGuard implements CanActivate {
   constructor(private auth: AuthService, private router: Router) { }
 
   canActivate(route, state: RouterStateSnapshot){
-    return this.auth.user$.map((user) => {
+    return this.auth.user$.pipe(map((user) => {
       if(user && user.isAdmin) return true;
       this.router.navigate(['access-denied']);;
-    });
+    }));
   }
 }
